Text.Toml

Description

Parsing and traversing TOML, some guy's weird replacement for .ini files.

Parsing TOML

The parser

toml :: (TokenParsing m, MonadPlus m) => m Toml

Parse some TOML.

The TOML AST

type Toml = Map Text Value

Represents TOML.

data Value

Top-level TOML values.

# dictionary
[some_dictionary]
val1 = "foo"

# list of dictionaries
[[some_item]]
val2 = "bar"

[[some_item]]
val3 = "baz"

Constructors

Table Toml

Tables (Seq Toml)

Scalar Scalar

data Scalar

TOML scalar values; i.e., values that can be associated with keys.

str = "bar"
decimal = 1
floating = 1.0
date = 1999-12-31T23:59:59Z
boolean = true
list = [ 1, 2, 3, 4, 5 ]

Constructors

String Text

Decimal Integer

Floating Double

Bool Bool

Date UTCTime

List (Vector Scalar)

Lenses for TOML

Traversing anything JSON-like in Haskell is usually a huge pain unless you can conceive some kind of convenient DSL for it. In this case we use utilities provided by the lens package.

Here's an example document; it's similar to the example on GitHub, with the exception that the 'servers' table is now a list. This more easily showcases some of the possibilities with the given lenses:

# This is a TOML document. Boom.

title = "TOML Example"

[owner]
name = "Tom Preston-Werner"
organization = "GitHub"
bio = "GitHub Cofounder & CEO\nLikes tater tots and beer."
dob = 1979-05-27T07:32:00Z # First class dates? Why not?

[database]
server = "192.168.1.1"
ports = [ 8001, 8001, 8002 ]
connection_max = 5000
enabled = true

[[servers]]
name = "alpha"
ip = "10.0.0.1"
dc = "eqdc10"

[[servers]]
name = "beta"
ip = "10.0.0.2"
dc = "eqdc10"

[clients]
data = [ ["gamma", "delta"], [1, 2] ]

# Line breaks are OK when inside arrays
hosts = [
  "alpha",
  "omega"
]

You can easily parse it using toml:

>>> Just ast <- parseFromFile toml "example.txt"

Using the functions that lens provides and the combinators here, you can then traverse deeply into the resultant structure to retrieve data:

>>> ast ^? table "owner" . string "bio" -- find Tom's bio
Just "GitHub Cofounder & CEO\nLikes tater tots and beer."

>>> ast ^.. tables "servers" . string "ip" -- list every server IP
["10.0.0.1","10.0.0.2"]

>>> ast ^.. table "database" . listOf "ports" _Decimal -- get database ports
[8001,8001,8002]

You can even modify the AST:

>>> let newHosts = ast & table "clients" . listOf "hosts" _String %~ T.reverse
>>> newHosts ^? table "clients" . list "hosts"
Just (fromList [String "ahpla",String "agemo"])

Top-level value lenses

table :: Text -> Traversal' Toml Toml

Traverse a TOML table, given a key name.

>>> "[foo]" ^? table "foo"
Just (fromList [])

tables :: Text -> Traversal' Toml Toml

Traverse a list of TOML tables, given their shared key name.

>>> "[[foo]]\n[[foo]]" & lengthOf (tables "foo")
2

scalar :: Text -> Traversal' Toml Scalar

Target a generic Scalar value.

>>> "foo = \"bar\"" ^? scalar "foo"
Just (String "bar")

Scalar lenses

string :: Text -> Traversal' Toml Text

Given a key name, target a String value. Will fail if the key is present, but is not a String.

>>> "foo = \"bar\"" ^? string "foo"
Just "bar"

decimal :: Text -> Traversal' Toml Integer

Given a key name, target a Decimal value. Will fail if the key is present, but is not a Decimal.

>>> "foo = 1234" ^? decimal "foo"
Just 1234

floating :: Text -> Traversal' Toml Double

Given a key name, target a Floating value. Will fail if the key is present, but is not Floating.

>>> "foo = 1234.567" ^? floating "foo"
Just 1234.567

bool :: Text -> Traversal' Toml Bool

Given a key name, target a Bool value. Will fail if the key is present, but is not a Bool.

>>> "foo = false" ^? bool "foo"
Just False

date :: Text -> Traversal' Toml UTCTime

Given a key name, target a Date value. Will fail if the key is present, but is not a Date.

>>> "foo = 1994-04-28T05:30:22Z" ^? date "foo"
Just 1994-04-28 05:30:22 UTC

list :: Text -> Traversal' Toml (Vector Scalar)

Given a key name, target a List value. Will fail if the key is present, but is not a List.

>>> "foo = [1, 2, 3]" ^? list "foo"
Just (fromList [Decimal 1,Decimal 2,Decimal 3])

listOf :: Text -> Prism' Scalar p -> Traversal' Toml p

Given a key name and a prism, target all the elements of a List that match the prism. Will fail if the key is present, but is not a List. Will fail if the list elements do not match the prism. (In TOML, lists are homogeneous, so you will always either retrieve the whole list or none of it.)

>>> "foo = [1, 2, 3]" ^.. listOf "foo" _Decimal
[1,2,3]

Top-level value prisms

_Table :: Prism' Value Toml

Focus on a table

_Tables :: Prism' Value (Seq Toml)

Focus on a list of tables

_Scalar :: Prism' Value Scalar

Focus on a scalar

Scalar prisms

_String :: Prism' Scalar Text

Focus on a String

_Decimal :: Prism' Scalar Integer

Focus on a Decimal

_Floating :: Prism' Scalar Double

Focus on a Floating

_Bool :: Prism' Scalar Bool

Focus on a Bool

_Date :: Prism' Scalar UTCTime

Focus on a Date

_List :: Prism' Scalar (Vector Scalar)

Focus on a List