Text.Toml.Pretty

Description

Pretty-print TOML.

Documentation

ppr :: Toml -> Doc

Pretty-prints TOML as JSON.

Pretty printing this example document:

# This is a TOML document. Boom.

title = "TOML Example"

[owner]
name = "Tom Preston-Werner"
organization = "GitHub"
bio = "GitHub Cofounder & CEO\nLikes tater tots and beer."
dob = 1979-05-27T07:32:00Z # First class dates? Why not?

[database]
server = "192.168.1.1"
ports = [ 8001, 8001, 8002 ]
connection_max = 5000
enabled = true

[[servers]]
name = "alpha"
ip = "10.0.0.1"
dc = "eqdc10"

[[servers]]
name = "beta"
ip = "10.0.0.2"
dc = "eqdc10"

[clients]
data = [ ["gamma", "delta"], [1, 2] ]

# Line breaks are OK when inside arrays
hosts = [
  "alpha",
  "omega"
]

yields the JSON:

{
  "clients": {
    "data": [["gamma", "delta"], [1, 2]],
    "hosts": ["alpha", "omega"]
  },
  "database": {
    "connection_max": 5000,
    "enabled": true,
    "ports": [8001, 8001, 8002],
    "server": "192.168.1.1"
  },
  "owner": {
    "bio": "GitHub Cofounder & CEO\nLikes tater tots and beer.",
    "dob": "1979-05-27T07:32:00Z",
    "name": "Tom Preston-Werner",
    "organization": "GitHub"
  },
  "servers": [
    {
      "dc": "eqdc10",
      "ip": "10.0.0.1",
      "name": "alpha"
    },
    {
      "dc": "eqdc10",
      "ip": "10.0.0.2",
      "name": "beta"
    }
  ],
  "title": "TOML Example"
}

(as a consequence of the internal structure of Map, the keys will always be in alphabetical order.)